requirements.txt
numpy
torch
torchvision
opencv_python
loguru
scikit-image
tqdm
Pillow
thop
ninja
tabulate
tensorboard
lap
filterpy
h5py
wandb
cython
pycocotools
cython_bbox
yacs
fvcore
progress
pandas
seaborn
